+++ src/config.rs
@@ -1,4 +1,5 @@
 pub(crate) static DEFAULT_N_TOP_CANDIDATES: usize = 5;
+pub(crate) static DEFAULT_CHAR_THRESHOLD: usize = 500;
 
 /// Configuration options for [`crate::Readability`]
 pub struct Config {
@@ -12,6 +13,8 @@ pub struct Config {
     pub disable_json_ld: bool,
     /// Number of top candidates to handle
     pub n_top_candidates: usize,
+    /// Character threshold for content extraction
+    pub char_threshold: usize,
 }
 
 
@@ -23,6 +26,7 @@ impl Default for Config {
             max_elements_to_parse: 0,
             disable_json_ld: false,
             n_top_candidates: DEFAULT_N_TOP_CANDIDATES,
+            char_threshold: DEFAULT_CHAR_THRESHOLD,
         }
     }
 }