+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to the `dom_smoothie` crate will be documented in this file.
 
 - Implement support for `Config::max_elements_to_parse` in `Readability::parse`.
 - Implement support for `Config::disable_json_ld` in `Readability::parse`.
+- Implement support for `Config::n_top_candidates` in `Readability::parse`.
+- Implement support for `Config::char_threshold` in `Readability::parse`.
+
 
 ### Fixed